Widen collector table columns to match DMV documentation
Full audit of all collector tables against Microsoft DMV docs found
6 SQL Server columns and 2 Lite columns narrower than documented types:

SQL Server:
- cpu_scheduler_stats: total_work_queue_count, total_active_parallel_thread_count (int -> bigint)
- cpu_scheduler_stats: system_memory_state_desc (nvarchar(120) -> nvarchar(256))
- waiting_tasks: resource_description (nvarchar(1000) -> nvarchar(3072))
- database_size_stats: recovery_model_desc (nvarchar(12) -> nvarchar(60))
- database_size_stats: volume_mount_point (nvarchar(256) -> nvarchar(512))

Lite (DuckDB):
- tempdb_stats: total_sessions_using_tempdb (INTEGER -> BIGINT, removed narrowing cast)
- session_stats: connection_count (INTEGER -> BIGINT, COUNT_BIG result)

Also fixes Azure SQL DB collector CONVERT(nvarchar(12),...) -> nvarchar(60)
for recovery_model_desc.

Idempotent upgrade script tested on sql2022.

+++ Lite/Database/Schema.cs
@@ -348,7 +348,7 @@ CREATE TABLE IF NOT EXISTS tempdb_stats (
     version_store_reserved_mb DECIMAL(18,2),
     total_reserved_mb DECIMAL(18,2),
     unallocated_mb DECIMAL(18,2),
-    total_sessions_using_tempdb INTEGER,
+    total_sessions_using_tempdb BIGINT,
     top_session_id INTEGER,
     top_session_tempdb_mb DECIMAL(18,2)
 )";
@@ -652,7 +652,7 @@ CREATE TABLE IF NOT EXISTS session_stats (
     server_id INTEGER NOT NULL,
     server_name VARCHAR NOT NULL,
     program_name VARCHAR NOT NULL,
-    connection_count INTEGER NOT NULL,
+    connection_count BIGINT NOT NULL,
     running_count INTEGER NOT NULL,
     sleeping_count INTEGER NOT NULL,
     dormant_count INTEGER NOT NULL,
